+++ src/web/FloraWeb/Components/PackageListHeader.hs
@@ -15,10 +15,9 @@ presentationHeader
   -> FloraHTML
 presentationHeader title subtitle numberOfPackages = do
   div_ [class_ "divider"] $ do
-    div_ [class_ "page-title"] $
+    div_ [class_ "page-title"] $ do
       h1_ [class_ ""] $ do
         span_ [class_ "headline"] $ toHtml title
-        toHtmlRaw @Text "&nbsp;"
-        span_ [class_ "dark:text-gray-200 version"] $ toHtml $ display numberOfPackages <> " results"
+    p_ [class_ "package-count"] $ toHtml $ display numberOfPackages <> " results"
     div_ [class_ "synopsis lg:text-xl text-center"] $
       p_ [class_ ""] (toHtml subtitle)